Refaktoroi ensimmäiset osat koodisto:sta common-packageen
