+++ src/main/scala/fi/oph/common/koodisto/Koodisto.scala
@@ -1,4 +1,4 @@
-package fi.oph.koski.koodisto
+package fi.oph.common.koodisto
 
 import java.time.LocalDate
 
@@ -8,7 +8,9 @@ import fi.oph.scalaschema.annotation.DefaultValue
 // Tätä oliota käytetään myös koodistojen luonnissa. Älä poista kenttiä!
 case class Koodisto(koodistoUri: String, versio: Int, metadata: List[KoodistoMetadata], codesGroupUri: String, voimassaAlkuPvm: LocalDate, organisaatioOid: String, withinCodes: Option[List[KoodistoRelationship]] = None, @DefaultValue("LUONNOS") tila: String = "LUONNOS", @DefaultValue(0) version: Int = 0) {
   def koodistoViite = KoodistoViite(koodistoUri, versio)
+
   def nimi = localizedStringFromMetadata { meta => meta.nimi.map(_.trim) }
+
   def kuvaus = localizedStringFromMetadata { meta => meta.kuvaus.map(_.trim) }
 
   private def localizedStringFromMetadata(f: KoodistoMetadata => Option[String]): Option[LocalizedString] = {
