+++ src/main/scala/fi/oph/koski/documentation/KoodistoServlet.scala
@@ -1,8 +1,9 @@
 package fi.oph.koski.documentation
 
+import fi.oph.common.koodisto.{Koodisto, KoodistoKoodi, KoodistoViite}
 import fi.oph.koski.config.KoskiApplication
 import fi.oph.koski.http.KoskiErrorCategory
-import fi.oph.koski.koodisto.{Koodisto, KoodistoKoodi, KoodistoPalvelu, KoodistoViite}
+import fi.oph.koski.koodisto.KoodistoPalvelu
 import fi.oph.koski.koskiuser.Unauthenticated
 import fi.oph.koski.schema.Opiskeluoikeus
 import fi.oph.koski.servlet.{ApiServlet, KoskiBaseServlet, NoCache}
