+++ src/main/scala/fi/oph/koski/editor/EditorKooditServlet.scala
@@ -1,10 +1,10 @@
 package fi.oph.koski.editor
 
+import fi.oph.common.koodisto.KoodistoViite
 import fi.oph.koski.config.KoskiApplication
 import fi.oph.koski.http.KoskiErrorCategory
 import fi.oph.koski.http.KoskiErrorCategory.badRequest.validation.koodisto.tuntematonKoodi
 import fi.oph.koski.json.LegacyJsonSerialization
-import fi.oph.koski.koodisto.KoodistoViite
 import fi.oph.koski.koskiuser.RequiresVirkailijaOrPalvelukäyttäjä
 import fi.oph.koski.schema._
 import fi.oph.koski.servlet.{ApiServlet, NoCache}
