+++ src/main/scala/fi/oph/koski/http/HttpStatusException.scala
@@ -1,7 +1,7 @@
 package fi.oph.koski.http
 
+import fi.oph.common.log.Loggable
 import fi.oph.koski.log.LogUtils.maskSensitiveInformation
-import fi.oph.koski.log.Loggable
 import org.http4s.Request
 
 /**
