+++ src/main/scala/fi/oph/koski/koodisto/KoodistoCreator.scala
@@ -2,11 +2,12 @@ package fi.oph.koski.koodisto
 
 import java.time.LocalDate
 
+import fi.oph.common.koodisto.{Koodisto, KoodistoKoodi, KoodistoViite}
 import fi.oph.koski.config.KoskiApplication
 import fi.oph.koski.json.JsonDiff.objectDiff
 import fi.oph.koski.json.JsonSerializer
 import fi.oph.koski.log.Logging
-import fi.oph.koski.koodisto.MockKoodistoPalvelu.{sortKoodistoMetadata, sortKoodistoKoodiMetadata}
+import fi.oph.koski.koodisto.MockKoodistoPalvelu.{sortKoodistoKoodiMetadata, sortKoodistoMetadata}
 import org.json4s.jackson.JsonMethods
 
 case class KoodistoCreator(application: KoskiApplication) extends Logging {
