+++ src/main/scala/fi/oph/koski/koodisto/KoodistoMockDataSorter.scala
@@ -1,5 +1,6 @@
 package fi.oph.koski.koodisto
 
+import fi.oph.common.koodisto.{Koodisto, KoodistoKoodi}
 import fi.oph.koski.json.{JsonFiles, JsonSerializer}
 import fi.oph.koski.log.Logging
