+++ src/main/scala/fi/oph/koski/koodisto/KoodistoMockDataUpdater.scala
@@ -1,10 +1,11 @@
 package fi.oph.koski.koodisto
 
 import com.typesafe.config.Config
+import fi.oph.common.koodisto.KoodistoKoodi
 import fi.oph.koski.config.KoskiApplication
 import fi.oph.koski.json.JsonFiles
 import fi.oph.koski.log.Logging
-import fi.oph.koski.koodisto.MockKoodistoPalvelu.{sortKoodistoMetadata, sortKoodistoKoodiMetadata}
+import fi.oph.koski.koodisto.MockKoodistoPalvelu.{sortKoodistoKoodiMetadata, sortKoodistoMetadata}
 
 object KoodistoMockDataUpdater extends App with Logging {
   updateMockDataFromKoodistoPalvelu(KoskiApplication.defaultConfig)
