+++ src/main/scala/fi/oph/koski/koodisto/KoodistoViitePalvelu.scala
@@ -1,10 +1,12 @@
 package fi.oph.koski.koodisto
 
+import fi.oph.common.koodisto.{KoodistoKoodi, KoodistoViite}
 import fi.oph.koski.cache._
 import fi.oph.koski.http.KoskiErrorCategory
 import fi.oph.koski.log.Logging
 import fi.oph.koski.schema.Koodistokoodiviite
 import fi.oph.koski.servlet.InvalidRequestException
+
 import scala.concurrent.duration._
 
 case class KoodistoViitePalvelu(koodistoPalvelu: KoodistoPalvelu)(implicit cacheInvalidator: CacheManager) extends Logging {
