+++ src/main/scala/fi/oph/koski/koodisto/MockKoodistoPalvelu.scala
@@ -1,4 +1,5 @@
 package fi.oph.koski.koodisto
+import fi.oph.common.koodisto.{Koodisto, KoodistoKoodi, KoodistoViite}
 import fi.oph.koski.cache.{Cached, GlobalCacheManager}
 import fi.oph.koski.json.{JsonResources, JsonSerializer}
 import fi.oph.koski.koodisto.MockKoodistoPalvelu._
