+++ src/main/scala/fi/oph/koski/raportointikanta/KoodistoLoader.scala
@@ -1,6 +1,7 @@
 package fi.oph.koski.raportointikanta
 
-import fi.oph.koski.koodisto.{KoodistoKoodi, KoodistoPalvelu}
+import fi.oph.common.koodisto.KoodistoKoodi
+import fi.oph.koski.koodisto.KoodistoPalvelu
 import fi.oph.koski.log.Logging
 import fi.oph.koski.raportointikanta.LoaderUtils.convertLocalizedString
