+++ src/main/scala/fi/oph/koski/virta/VirtaClient.scala
@@ -1,10 +1,11 @@
 package fi.oph.koski.virta
 
 import com.typesafe.config.Config
+import fi.oph.common.log.NotLoggable
 import fi.oph.koski.config.{Environment, SecretsManager}
 import fi.oph.koski.http.Http._
 import fi.oph.koski.http.{Http, HttpConnectionException}
-import fi.oph.koski.log.{Logging, NotLoggable, TimedProxy}
+import fi.oph.koski.log.{Logging, TimedProxy}
 import fi.oph.koski.util.Files
 
 import scala.xml.{Elem, Node}
