+++ src/main/scala/fi/oph/koski/virta/VirtaOpiskeluoikeusRepository.scala
@@ -1,12 +1,12 @@
 package fi.oph.koski.virta
 
+import fi.oph.common.log.NotLoggable
 import fi.oph.koski.cache.{CacheManager, GlobalCacheManager}
 import fi.oph.koski.config.KoskiApplication
 import fi.oph.koski.henkilo.HenkilönTunnisteet
 import fi.oph.koski.http.HttpStatus
 import fi.oph.koski.koodisto.{KoodistoViitePalvelu, MockKoodistoViitePalvelu}
 import fi.oph.koski.koskiuser.{AccessChecker, AccessType, KoskiSession, SkipAccessCheck}
-import fi.oph.koski.log.NotLoggable
 import fi.oph.koski.opiskeluoikeus.AuxiliaryOpiskeluoikeusRepositoryImpl
 import fi.oph.koski.oppilaitos.{MockOppilaitosRepository, OppilaitosRepository}
 import fi.oph.koski.organisaatio.{MockOrganisaatioRepository, OrganisaatioRepository}
