+++ src/main/scala/fi/oph/koski/ytr/YtrOpiskeluoikeusRepository.scala
@@ -1,12 +1,12 @@
 package fi.oph.koski.ytr
 
+import fi.oph.common.log.NotLoggable
 import fi.oph.koski.cache.CacheManager
 import fi.oph.koski.henkilo.HenkilönTunnisteet
 import fi.oph.koski.http.HttpStatus
 import fi.oph.koski.koodisto.KoodistoViitePalvelu
 import fi.oph.koski.koskiuser.{AccessChecker, AccessType, KoskiSession}
 import fi.oph.koski.localization.LocalizationRepository
-import fi.oph.koski.log.NotLoggable
 import fi.oph.koski.opiskeluoikeus.AuxiliaryOpiskeluoikeusRepositoryImpl
 import fi.oph.koski.oppilaitos.OppilaitosRepository
 import fi.oph.koski.organisaatio.OrganisaatioRepository
