+++ src/test/scala/fi/oph/koski/json/JsonDiffTest.scala
@@ -2,7 +2,8 @@ package fi.oph.koski.json
 
 import java.time.LocalDate
 
-import fi.oph.koski.koodisto.Koodisto
+import fi.oph.common.koodisto
+import fi.oph.common.koodisto.Koodisto
 import org.json4s.JsonAST.{JInt, JObject}
 import org.json4s.jackson.JsonMethods
 import org.scalatest.{FreeSpec, Matchers}
@@ -14,7 +15,7 @@ class JsonDiffTest extends FreeSpec with Matchers {
     JsonMethods.compact(JsonDiff.jsonDiff(a, b)) should equal("""[{"op":"replace","path":"/b","value":3}]""")
   }
   "objectDiff" in {
-    val a = Koodisto("a", 1, Nil, "group", LocalDate.MIN, "ord", None, "LUONNOS", 0)
+    val a = koodisto.Koodisto("a", 1, Nil, "group", LocalDate.MIN, "ord", None, "LUONNOS", 0)
     val b = a.copy(koodistoUri = "b")
     JsonMethods.compact(JsonDiff.objectDiff(a, b)) should equal("""[{"op":"replace","path":"/koodistoUri","value":"b"}]""")
   }
